changes in constans, and files
