+++ React/game/src/constans/form.js
@@ -1,0 +1,1 @@
+export const nameForm = 'loginForm';
