+++ React/game/src/constans/game.js
