--- React/game/src/services/BookService.js
@@ -1,5 +1,0 @@
-import api from '../../config/api';
-
-export default {
-  getBookDetail: id => api.get('/books', { id })
-};
